Refactor and deduplication
TODO - refactor unit tests

Changes:
- Removed redundant services and interfaces
- Removed redundant function
