--- NCS.DSS.GDPRUpdateCustomerInformation/Services/IIdentifyAndAnonymiseDataService.cs
@@ -1,9 +1,0 @@
-﻿namespace NCS.DSS.DataUtility.Services
-{
-    public interface IIdentifyAndAnonymiseDataService
-    {
-        Task AnonymiseData();
-        Task DeleteCustomersFromCosmos(List<Guid> customerIdList);
-        Task<List<Guid>> ReturnCustomerIds();
-    }
-}
